strands-examples/requirements.txt
python-dotenv==1.1.1
starlette==0.47.2
strands-agents[litellm]==1.0.1
strands-agents-tools==0.2.1
streamlit==1.47.1
